Test modifications and rethinking module graph/naming
