--- util.ts
@@ -1,1 +1,0 @@
-// Copyright 2022 Connor Logan. All rights reserved. MIT License.
